Implement maxArea method for water container problem using two pointer
This commit adds a solution for the LeetCode problem “Container With Most Water”. The maxArea function calculates the maximum amount of water a container can store, given a list of vertical line heights.

The approach uses two pointers:

i starts at the beginning of the list and j at the end. The width of the container is w = j - i.
The height of the container is the smaller of the two lines: h = min(height[i], height[j]).
The area is calculated as area = w * h.
maxwater is updated if this area is larger than previous areas.
The pointer pointing to the smaller height is moved inward to potentially find a larger area.
This process continues until the two pointers meet.
Time Complexity: O(n) – each element is visited at most once.
Space Complexity: O(1) – no additional space is used beyond a few variables.